backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.68.0,<0.69.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "psycopg2-binary>=2.9.1,<3.0.0",
    "pydantic>=1.8.0,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "python-multipart>=0.0.5,<0.1.0",
    "spacy>=3.5.0,<4.0.0",
    "sqlalchemy>=1.4.23,<1.5.0",
    "uvicorn>=0.15.0,<0.16.0",
]
